backend/requirements.txt
fastapi >= 0.111.1, < 0.112
orjson >= 3.10.6, < 4
pydantic-settings >= 2.3.4, < 3
fastapi-cli >= 0.0.4
sqlmodel >= 0.0.20, < 0.0.21
alembic >= 1.13.2, < 1.14
slowapi >= 0.1.9, < 0.2
requests >= 2.32.3, < 3
python-dateutil >= 2.9.0.post0, < 3
fastapi-cache2 >= 0.2.1, < 0.3
msgpack-asgi==1.*
sentry-sdk[fastapi]
python-slugify >= 8.0.4, < 9